pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ood-solver"
version = "0.1.0"
description = "Phase-1 structured OOD solver prototype"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "matplotlib>=3.8",
  "tqdm>=4.66",
]
